% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{AVal}
\alias{AVal}
\title{Calculate A-value}
\usage{
AVal(x.sample, y.ref, A.method)
}
\arguments{
\item{x.sample}{A vector of normalized \eqn{log_2} intensity of tissue
sample.}

\item{y.ref}{A vector of normalized \eqn{log_2} intensity of reference
sample.}

\item{A.method}{A character string indicating the method to calculate
A-value. One of \code{"mean_of_2samples"}, \code{"ref"} or \code{"sample"}.
If it is \code{"mean_of_2samples"}, A-value is the mean of tissue sample
and reference sample; else if it is \code{"ref"}, A-value take as the
intensity of reference sample; else if it is \code{"sample"}, A-value take
as the intensity of tissue sample. \code{"ref"} is recommended.}
}
\value{
A vector of A-value.
}
\description{
Calculate A-value
}
