% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobalPlot.R
\name{barplot_NLC_R2}
\alias{barplot_NLC_R2}
\title{Plot \eqn{R^2} of nonlinear model fitting for all samples}
\usage{
barplot_NLC_R2(
  nonlinearR2,
  para = c("var", "bias"),
  para.fit.method,
  main = "",
  outdir
)
}
\arguments{
\item{nonlinearR2}{A vector of all \eqn{R^2} of nonlinear models.}

\item{para}{A character indicates which parameter to be estimated in the
nonlinear models.}

\item{para.fit.method}{A character string indicating the method to be used
for nonlinear model fitting globally. Either \code{"natural_cubic_spline"}
or \code{"log_trans"} (default) or \code{"local_regression"}.}

\item{main}{The title of the plot, usually it is the name of the
\code{\link{proObj}}.}

\item{outdir}{The out put directory.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Plot \eqn{R^2} of nonlinear model fitting for all samples
}
