% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQC.R
\name{dataQC}
\alias{dataQC}
\title{Quality control of raw intensity}
\usage{
dataQC(x, outdir = "./", outadd = "", cpus = 5)
}
\arguments{
\item{x}{A \code{\link{proObj}}.}

\item{outdir}{The output directory.}

\item{outadd}{A character string to describe the additional information of
plots.}

\item{cpus}{The number of CPUs requested for a parallel computation.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Given the raw intensity of all tissue samples, plot the number of detected
samples for each protein and the number of detected proteins in each sample.
}
