% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffVarTest.R
\name{DiffVar}
\alias{DiffVar}
\title{Identify differentially variable proteins}
\usage{
DiffVar(
  x,
  y,
  scaleV.raw.FC = 0.1,
  alternative = c("upper", "lower", "two.sided"),
  p.adjust.method = "BH"
)
}
\arguments{
\item{x, y}{\code{x} is any R object for which a \code{diffTest} method has
been defined. For the method for class \code{\link{proObj}}, \code{x} and
\code{y} are two \code{proObj} objects to be compared. They must be
associated with estimated \emph{v}, Fisher information and scaled by the
same baseline \eqn{v_0}.}

\item{scaleV.raw.FC}{A specified average basic fluctuation around mean
expression of a protein. The default setting of \code{scaleV.raw.FC} is
10\%, which amounts to an average fluctuation of 10\% around mean expression.
Accordingly, the baseline fluctuation for each protein is defined as
\eqn{log(log_2(1+10\%)^2)}. If \code{scaleV.raw.FC} is 0, the differential
variability test would be conducted without scaling of estimated \emph{v}.}

\item{alternative}{The alternative hypothesis. Must be one of \code{"upper"}
(default), \code{"lower"}, and \code{"two.sided"}. Can be abbreviated.}

\item{p.adjust.method}{The method for adjusting \emph{p}-values for multiple
testing. See \code{\link[stats]{p.adjust}} for all options.}
}
\value{
An object of the \code{\link[base]{class}} \code{c("diffVar",
  "data.frame")}, which records the differential variability test results for
  each protein by each row.
}
\description{
Conduct hypothesis test for differentially variable proteins. All
\code{\link{proObj}} objects should be associated with the same scaling of
estimated \emph{v}. Thus, they must have the same scaling results of
\emph{v}.
}
\details{
Proteins which are located in the lower part of \emph{v} are obviously
non-DVP, their \emph{p}-values are set as \code{NA} directly.
}
