% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffVarTest.R
\name{diffVarPlot}
\alias{diffVarPlot}
\title{Plot a \code{diffVar} Object}
\usage{
diffVarPlot(
  res,
  ref.mean,
  P.cutoff = 0.05,
  p.type = c("pval", "padj"),
  Nsample,
  sample.type1,
  sample.type2,
  args.points = list(pch = 21, cex = 1.2),
  args.abline = list(col = "darkorange", lwd = 2.5, lty = 5),
  args.legend = list(x = "topright"),
  xlab = "mean",
  ylab = "v = log(Var)",
  xlim = NULL,
  ylim1 = NULL,
  ylim2 = NULL,
  outdir = "./",
  outadd = "",
  ...
)
}
\arguments{
\item{res}{A \code{diffVar} object.}

\item{ref.mean}{A numeric vector of mean among all reference samples.}

\item{P.cutoff}{A scalar between 0 and 1. Proteins with \emph{p}-values below
this cutoff are considered statistically significant under the null
hypothesis and are correspondingly highlighted in a distinct color in the
plot.}

\item{p.type}{A character which indicates the type of \emph{p}-value to show
in the plot. Must be one of \code{"pval"} (default), and \code{"padj"}. Can
be abbreviated.}

\item{Nsample}{The sample size in this group.}

\item{sample.type1}{The sample type1 of differential variability test.}

\item{sample.type2}{The sample type2 of differential variability test.}

\item{args.points}{A named list of graphical parameters for plotting the
point estimates (see \code{\link[graphics]{points}} for commonly used
graphical parameters).}

\item{args.abline}{A named list of graphical parameters for plotting the null
\emph{v_0} (see \code{\link[graphics]{abline}} for commonly used graphical
parameters).}

\item{args.legend}{A named list of graphical parameters for legend (see
\code{\link[graphics]{legend}} for commonly used graphical parameters).}

\item{xlab}{Label for the X axis.}

\item{ylab}{Label for the Y axis.}

\item{xlim}{Range of X coordinate.}

\item{ylim1}{Range of Y coordinate in subplot1.}

\item{ylim2}{Range of Y coordinate in subplot2.}

\item{outdir}{If \code{PLOT} is TRUE, the out put direction.}

\item{outadd}{If \code{PLOT} is TRUE, the additions to the out put documents.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}}.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Given a \code{diffVar} object, which records the results of calling
HVRs and/or LVRs from a \code{\link{proObj}}, this method creates a scatter
plot of \code{(ref.mean, v)} pairs from all observations, marking
specifically the ones that are associated with statistical significance.
Besides, the theoretical \eqn{v_0} associated with this round of test of
\eqn{v} is added to the plot, serving as a baseline to which each sample
variance can be compared.
}
\seealso{
\code{\link{proObj}} for creating a \code{proObj} object;
  \code{\link{estTechVar}} for fitting technical variance given a
  \code{proObj} objects; \code{\link{estBioVar}} for estimating biological
  mean \emph{m} and \emph{log}-transformed biological variance \emph{v};
  \code{diffVar} for a test of differential-variability.
}
