% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobalPlot.R
\name{est_para_plot}
\alias{est_para_plot}
\title{Plot the final estimated global variance or bias curve}
\usage{
est_para_plot(
  AValue,
  estPara,
  R2,
  outdir,
  sample,
  para = c("var", "bias"),
  spc_info = "",
  outname = "nonlinear_model_fitted"
)
}
\arguments{
\item{AValue}{A vector of A-value. \code{AValue} would be used to predict the
corresponding estimated variance from the curve.}

\item{estPara}{A numeric vector of final estimated parameters.}

\item{R2}{The goodness of fit \eqn{R^2} of the nonlinear fitting}

\item{outdir}{The output directory of the plot.}

\item{sample}{The name of current sample to be fitted.}

\item{para}{A character string to clarify which parameter would be fitted in
this round. Either \code{"var"} or \code{"bias"}.}

\item{spc_info}{A character string containing the specific path to output the
process plots of current nonlinear model.}

\item{outname}{A character string to describe more about which kind of
nonlinear model was used for fitting.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Although nonlinear curves are obtained from \code{fit_natural_cubic_model},
\code{fit_log_trans_model} or \code{fit_local_model}, more judgement need to
be done to decide whether use estimation from the curve directly. The final
estimation curve may not be strictly the same as the fitted curve.
\code{est_para_plot} would plot the final estimation curves.
}
