% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobal.R
\name{estTechParaGlobal_Each}
\alias{estTechParaGlobal_Each}
\title{Globally fit the technical variance and bias for current sample}
\usage{
estTechParaGlobal_Each(
  resTechLocal,
  para = c("var", "bias"),
  para.fit.method,
  used.global.fit,
  PLOT = FALSE,
  outdir
)
}
\arguments{
\item{resTechLocal}{A list which includes all the locally fitted technical
results.}

\item{para}{A character string to clarify which parameter would be fitted in
this round. Either \code{"var"} or \code{"bias"}.}

\item{para.fit.method}{A character string indicating the method to be used
for fitting the nonlinear curve globally. When \code{para} is \code{"var"},
\code{para.fit.method} should be one of \code{"natural_cubic_spline"},
\code{"log_trans"} or \code{"local_regression"}; and when \code{para} is
\code{"bias"}, \code{para.fit.method} should be either
\code{"natural_cubic_spline"} or \code{"local_regression"}}

\item{used.global.fit}{A vector of 2 with each element between 0 and 1
indicates which parts of sliding windows could be used for globally curve
fitting. Since the locally estimated parameters in first and end parts of
the sliding window may not be stable and unreliable, \code{used.global.fit}
can specify that only this part of the sliding windows are used. The
default is \eqn{[0.05, 0.95]}.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}
}
\value{
A list include globally estimated technical curves and the
  corresponding predicted parameters.
}
\description{
By using the locally fitted technical parameters \code{resTechLocal}, fit the
technical variance and normalization bias globally for current sample. The
nonlinear fitted curve is a function of \code{AValue}.
}
