% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobal.R
\name{estTechVarGlobal}
\alias{estTechVarGlobal}
\title{Globally fit the technical variance and bias for each sample}
\usage{
estTechVarGlobal(
  x,
  var.fit.method = "log_trans",
  bias.fit.method = "log_trans",
  used.global.fit = c(0.05, 0.95),
  correct.M = T,
  PLOT = F,
  outdir = "./"
)
}
\arguments{
\item{x}{A \code{\link{proObj}} with an available \code{resTechLocal} field
in its \code{resTech} field.}

\item{var.fit.method}{A character string indicating the method to be used for
fitting variance curve globally. One of \code{"natural_cubic_spline"} or
\code{"log_trans"} (default) or \code{"local_regression"}.}

\item{bias.fit.method}{A character string indicating the method to be used
for fitting bias curve globally. Either \code{"natural_cubic_spline"}
(default) or \code{"local_regression"}.}

\item{used.global.fit}{A vector of 2 with each element between 0 and 1
indicates which parts of sliding windows could be used for globally curve
fitting. Since the locally estimated parameters in first and end parts of
the sliding window may not be stable and unreliable, \code{used.global.fit}
can specify that only this part of the sliding windows are used. The
default is \eqn{[0.05, 0.95]}.}

\item{correct.M}{Whether to correct M-value by fitted global bias curve?}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}
}
\value{
A \code{\link{proObj}} with globally estimated technical variance and
  normalization bias.

\code{estTechVarGlobal} returns a \code{\link{proObj}} object, it has
  has an added (updated) \code{resTechGlobal} in \code{resTech} field
  describing globally estimated technical variance and normalization bias. If
  \code{correct.M} is TRUE, the \code{\link{proObj}} object would have an
  added \code{adj.MValue} in addition. The \code{resTechGlobal} field is
  itself a list of all samples, and each one is consist of the following
  components:
    \describe{
        \item{\code{nlc.var}}{A list of all samples, each of them records the
         globally fitted technical variance prediction function of current
         sample.}
        \item{\code{nlc.bias}}{A list of all samples, each of them records the
         globally fitted normalization bias prediction function of current
         sample.}
        \item{\code{est.tech.var}}{A matrix whose number of rows is the same
        as number of all proteins and number of columns is the same as sample
        size. It indicates the final estimated technical variance.}
        \item{\code{est.bias}}{A matrix whose number of rows is the same
        as number of all proteins and number of columns is the same as sample
        size. It indicates the final estimated normalization bias.}
    }
  \code{adj.MValue} is a matrix whose number of rows is the same as number of
  all proteins and number of columns is the same as sample size.
  \code{MValue} is adjusted by normalization bias.
}
\description{
Given a \code{\link{proObj}}, this function globally fit the nonlinear model
for technical variance and bias in each sample by using the local results
stored in \code{resTechLocal} field.
}
