% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocal.R
\name{estTechVarLocal}
\alias{estTechVarLocal}
\title{Locally fit the technical variance and bias for each sample}
\usage{
estTechVarLocal(
  x,
  A.method = "ref",
  filter.low = T,
  offset = 1,
  window.size = 400,
  step.size = 100,
  middle.percent = 0.5,
  cpus = 1,
  PLOT = FALSE,
  outdir = "./"
)
}
\arguments{
\item{x}{A \code{\link{proObj}}.}

\item{A.method}{A character string indicating the method to calculate
A-value.}

\item{filter.low}{Whether to filter proteins whose raw intensity is under 10?
The default is TRUE.}

\item{offset}{A scalar used as an offset added to the raw intensity, avoiding
infinity while log-transformation. Default is 1.}

\item{window.size}{A number which indicates the window size of sliding window
when estimate technical variance and normalization bias locally.}

\item{step.size}{A number which indicates the step size of sliding window
when estimate technical variance and normalization bias locally.}

\item{middle.percent}{A number that specifies the percentage of parameters used
for linear model fitting within each sliding window.}

\item{cpus}{The number of CPUs requested for a parallel computation. By
default, no parallel computation is applied.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}
}
\value{
\code{estTechVarLocal} returns a \code{\link{proObj}} object, it has
  has an added (updated) \code{resTechLocal} in \code{resTech} field
  describing locally estimated technical variance and normalization bias, a
  \code{MValue} field recording M-value of all samples and a \code{ref.mean}
  field indicating mean of all reference samples. The \code{resTechLocal}
  field is itself a list of all samples, and each one is consist of the
  following components:
    \describe{
        \item{\code{log2.norm.int}}{A normalized intensity 3-column matrix
        with normalized intensity, M-value and A-value in the \eqn{log_2}
        scale.}
        \item{\code{MA.linear.fit}}{A 4-column matrix records all of the
        results about linear fitting in each sliding window of current
        sample, which includes fitted variances, bias, mean of A-value in
        each window and goodness of fit \eqn{R^2}.}
        \item{\code{name}}{The name of current sample.}
    }
  \code{MValue} is a matrix whose number of rows is the same as number of all
  proteins and number of columns is the same as sample size. It record
  M-value (i.e. \eqn{log_2}-transformed fold change) of all samples.

  \code{AValue} is a matrix whose number of rows is the same as \code{MValue}
  and the number of columns is the same as the number of reference samples.

  \code{ref.mean} is a vector whose length is the same as \code{MValue}. It
  indicates the mean of normalized \eqn{log_2}-intensity among all reference
  samples (i.e. mean of A-value among all reference samples), which is
  regarded as the basic intensity of the observations in the whole group of
  samples.
}
\description{
Given a \code{\link{proObj}}, this function locally fit the linear model for
technical variance and bias in each sample by linear fitting.
}
