% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocal.R
\name{estTechVarLocal_pair}
\alias{estTechVarLocal_pair}
\title{Normalize raw intensity and fit local parameters for each sample}
\usage{
estTechVarLocal_pair(
  sample,
  ref.sample,
  paired.int,
  A.method = "ref",
  filter.low = T,
  offset = 1,
  window.size,
  step.size,
  middle.percent,
  cpus,
  PLOT,
  outdir
)
}
\arguments{
\item{sample}{The name of current sample to be fitted.}

\item{ref.sample}{The name of reference sample related to current sample.}

\item{paired.int}{A matrix with 2 columns. The first column is the raw
intensity of current sample, and the second column is the raw intensity of
its corresponding reference sample.}

\item{A.method}{A character string indicating the method to calculate
A-value. One of \code{"mean_of_2samples"}, \code{"ref"} (default) or
\code{"sample"}. See \code{normalize} for more details.}

\item{filter.low}{Whether to filter proteins whose raw intensity is under 10?
The default is TRUE.}

\item{offset}{A scalar used as an offset added to the raw intensity, avoiding
infinity while log-transformation. Default is 1.}

\item{window.size}{A number which indicates the window size of sliding window
when estimate technical variance and normalization bias locally.}

\item{step.size}{A number which indicates the step size of sliding window
when estimate technical variance and normalization bias locally.}

\item{middle.percent}{A number that specifies the percentage of parameters used
for linear model fitting within each sliding window.}

\item{cpus}{The number of CPUs requested for a parallel computation. By
default, no parallel computation is applied.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}
}
\value{
A \code{list} records the normalized \eqn{log2}-intensity,
}
\description{
Normalize raw intensity and fit local parameters for each sample
}
