% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocal.R
\name{estTechVarLocal_pairFit}
\alias{estTechVarLocal_pairFit}
\title{Estimate local technical variance and normalization bias}
\usage{
estTechVarLocal_pairFit(
  log2.norm.int,
  sample,
  window.size = 400,
  step.size = 100,
  middle.percent = 0.5,
  PLOT = FALSE,
  cpus = 1,
  outdir = "./"
)
}
\arguments{
\item{log2.norm.int}{A matrix with normalized \eqn{log2}-intensity and the
corresponding A-value and M-value.}

\item{sample}{The name of current sample to be fitted.}

\item{window.size}{A number which indicates the window size of sliding window
when estimate technical variance and normalization bias locally.}

\item{step.size}{A number which indicates the step size of sliding window
when estimate technical variance and normalization bias locally.}

\item{middle.percent}{A number that specifies the percentage of parameters used
for linear model fitting within each sliding window.}

\item{PLOT}{Whether to draw the process plots?}

\item{cpus}{The number of CPUs requested for a parallel computation. By
default, no parallel computation is applied.}

\item{outdir}{If \code{PLOT} is true, the output directory.}
}
\value{
A matrix records all of the results about linear fitting in each
  sliding window, which includes fitted variances, bias, mean of A-value in
  each window and goodness of fit \eqn{R^2}.
}
\description{
Given normalized \eqn{log2}-intensity and the corresponding A-value and
M-value, scan the M-A plot by sliding window with \code{window.size} as
window size and \code{step.size} as step size. In each sliding window, plot
M-value against the corresponding quantile of standard normal distribution
and use a linear fitting model to clarify their relationship. The slope of
the linear model is local technical standard deviation of this window and the
intercept is the normalization bias.
}
