\name{EX_R}
\alias{EX_R}
\title{ Hamilton's FX Data }
\description{ 
  \code{EX_R} This is the French Franc/USD data was used in Hamilton and Engel's paper
  "Long Swings in the Dollar: Are They in the Data and Do Markets 
   Know It?", American Economic Review, Sept. 1990. 
}

\value{
  a list containing: 

  \item{EX_R}{Changes in French Franc/USD Exchange Rate}
}

\author{ James Eustace \email{james.eustace@omam.co.uk}}}
