% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{Fisher_Info}
\alias{Fisher_Info}
\title{Calculate Fisher information of \emph{m} and \emph{v}}
\usage{
Fisher_Info(bio.var, tech.var)
}
\arguments{
\item{bio.var}{A vector of biological variance for all proteins.}

\item{tech.var}{A matrix of technical variance for all proteins in all
samples.}
}
\value{
A matrix which records the Fisher information of biological mean
  \emph{m}, \eqn{log}-transformed biological variance \emph{v} and
  biological variance.
}
\description{
Calculate Fisher information of \emph{m} and \emph{v}
}
