% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarPlot.R
\name{FisherInfoPlot}
\alias{FisherInfoPlot}
\title{Plot Fisher information of \emph{m} and \emph{v}}
\usage{
FisherInfoPlot(
  m,
  v,
  fisher.m,
  fisher.v,
  Nsample,
  sample.type,
  group1,
  group2,
  scaleV.raw.FC = NULL,
  outdir = "",
  outadd
)
}
\arguments{
\item{m}{A numeric vector of biological mean estimates \emph{m} for all
proteins.}

\item{v}{A numeric vector of \eqn{log}-transformed biological variance
estimates \emph{v} for all proteins.}

\item{fisher.m}{A numeric vector of Fisher information of \emph{m}.}

\item{fisher.v}{A numeric vector of Fisher information of \emph{v}.}

\item{Nsample}{The sample size in this group.}

\item{sample.type}{The sample type of this group of samples.}

\item{group1, group2}{A vector indicating the first or second group of
proteins according to the classification of \emph{v}. If
\code{scaleV.raw.FC} is not \code{NULL}, the classification is conducted by
the baseline fluctuation \eqn{log(log_2^2(1+scaleV.raw.FC))}.}

\item{scaleV.raw.FC}{A specified average basic fluctuation around mean
expression of a protein. See \code{estBioVar} and \code{resBioCMB} for more
details.}

\item{outdir}{If \code{PLOT} is TRUE, the output directory.}

\item{outadd}{If \code{PLOT} is TRUE, the additions to the output documents.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
\code{FisherInfoPlot} would draw 4 subplots. The first 2 subplots in the
first row is the fisher information of \emph{m}, and proteins are divided in
2 groups by \code{group1} and \code{group2}. The other 2 subplots in the
second row is the fisher information of \emph{v}, groups as above.
}
