% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobalEngine.R
\name{fit_log_trans_model}
\alias{fit_log_trans_model}
\title{Fit global mean-para trend by \eqn{log}-transformed natural cubic spline}
\usage{
fit_log_trans_model(
  windowMean,
  windowPara,
  used.global.fit = c(0.05, 0.95),
  choice = 2,
  PLOT = FALSE,
  outdir,
  sample,
  para = c("var", "bias"),
  spc_info = "",
  outname = "log_trans_NCS_model"
)
}
\arguments{
\item{windowMean}{A vector of mean A-value in all sliding windows.}

\item{windowPara}{A vector of estimated local parameters in all sliding
windows.}

\item{used.global.fit}{A vector of 2 with each element between 0 and 1
specifying a subset of sliding windows as the input of the whole globally
curve fitting process. Since the locally estimated parameters in first and
end parts of the sliding windows may be unstable and unreliable,
\code{used.global.fit} can specify that only this part of the sliding
windows are used. The default is \code{c(0.05, 0.95)}.}

\item{choice}{A numeric to specify how many natural cubic spline nonlinear
models to fit to choose from. By default, 2 non linear models with
\code{df} in \code{\link[splines]{ns}} would be fitted and the one with
higher goodness of fit \eqn{R^2} would be served as the final nonlinear
model.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}

\item{sample}{The name of current sample to be fitted.}

\item{para}{A character string to clarify which parameter would be fitted in
this round. Either \code{"var"} or \code{"bias"}.}

\item{spc_info}{A character string containing the specific path to output the
process plots of current nonlinear model.}

\item{outname}{A character string to describe more about which kind of
nonlinear model was used for fitting.}
}
\value{
A prediction function which accepts a vector of means and returns the
  predicted parameters.
}
\description{
While fitting global mean-var curve, it should be ensured that all the fitted
variance curves are positive. \code{fit_log_trans_model} first transform all
\code{windowPara} in \eqn{log} scale, and then fit a mean-\eqn{log var} curve
by natural cubic spline. Finally it would return a re-scaled prediction
function in exponential level. This process can not only make sure that all
the data read out of the curve are positive, but also that the curve is
smoother because of the transformation.
}
\details{
\code{fit_log_trans_model} is more suitable to fit the technical variance
curve.
}
