% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalEngine.R
\name{getIndexes}
\alias{getIndexes}
\title{Calculate goodness of fit \eqn{R^2}}
\usage{
getIndexes(y.predict, y.data)
}
\arguments{
\item{y.predict}{A numeric vector of prediction.}

\item{y.data}{A numeric vector of observation.}
}
\value{
The goodness of fit \eqn{R^2}.
}
\description{
Calculate goodness of fit \eqn{R^2}
}
