\name{INT_D}
\alias{INT_D}
\title{ Hamilton's Interest Rate Data }
\description{ 
  \code{INT_D} This is the interest rate differential data was used in Hamilton and Engel's paper
  "Long Swings in the Dollar: Are They in the Data and Do Markets 
   Know It?", American Economic Review, Sept. 1990. 
}

\value{
  a list containing: 

  \item{INT_D}{Interest Rate Differential for France and United States}
}

\author{ James Eustace \email{james.eustace@omam.co.uk}}}
