% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{kmean_clus}
\alias{kmean_clus}
\title{Divide proteins into 2 classes according to \code{v}}
\usage{
kmean_clus(v, seed)
}
\arguments{
\item{v}{A numeric vector of all proteins' estimated \eqn{log}-biological
variance \emph{v}.}

\item{seed}{A single value, used to ensure the reproducibility of K-means
clustering result. Passed to \code{\link[base]{set.seed}}}
}
\value{
Labels of two classes of \code{v}.
}
\description{
Carry out K-means clustering on \code{v} and divide them into 2 groups.
Besides, check the labels of the observations to make sure the cluster with
higher cluster center is class 1 and the lower is class 2.
}
