% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalPlot.R
\name{linear_regression_function}
\alias{linear_regression_function}
\title{The expression of linear regression function}
\usage{
linear_regression_function(coef)
}
\arguments{
\item{coef}{The coefficients of the fitted linear model.}
}
\value{
The expression of linear model.
}
\description{
The expression of linear regression function
}
