% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalPlot.R
\name{local_linear_fitting_plot}
\alias{local_linear_fitting_plot}
\title{Draw the linear fitting plot in each sliding window}
\usage{
local_linear_fitting_plot(
  windowM,
  outdir,
  sample,
  window.Num,
  fit.region,
  theoretical.quantiles,
  R2,
  coef
)
}
\arguments{
\item{windowM}{A vector of M-value in current sliding window.}

\item{outdir}{A character which represent the output directory.}

\item{sample}{The name of current sample to be fitted.}

\item{window.Num}{The current number of sliding window.}

\item{fit.region}{The region of proteins to be used for fitting the linear
model.}

\item{theoretical.quantiles}{A vector of theoretical quantile correspond to
\code{windowM}.}

\item{R2}{The goodness of fit \eqn{R^2} of linear model.}

\item{coef}{The coefficients of linear model.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
\code{local_linear_fitting_plot} draw the linear fitting plot of all
proteins. The X-axis demonstrates theoretical quantiles of standard normal
distribution. The Y-axis demonstrates M-values in current window. Proteins
used for linear fitting are in \code{"red"} and others are in \code{"grey"}.
The fitted linear model is also added in the plot.
}
