% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobalPlot.R
\name{log_trans_NCS_plot}
\alias{log_trans_NCS_plot}
\title{Plot mean-parameter in all sliding windows and fitted curve from
\eqn{log}-transformed natural cubic spline model}
\usage{
log_trans_NCS_plot(
  windowMean,
  windowPara,
  outdir,
  sample,
  para = c("var", "bias"),
  spc_info = "",
  outname = "log_trans_NCS_model",
  fit.region,
  fit,
  R2,
  df
)
}
\arguments{
\item{windowMean}{A vector of mean A-value in all sliding windows.}

\item{windowPara}{A vector of estimated local parameters in all sliding
windows.}

\item{outdir}{The output directory of the plot.}

\item{sample}{The name of current sample to be fitted.}

\item{para}{A character string to clarify which parameter would be fitted in
this round. Either \code{"var"} or \code{"bias"}.}

\item{spc_info}{A character string containing the specific path to output the
process plots of current nonlinear model.}

\item{outname}{A character string to describe more about which kind of
nonlinear model was used for fitting.}

\item{fit.region}{A vector of 2 specifying a subset of sliding windows as the
input of the whole globally curve fitting process.}

\item{fit}{The prediction function which accepts a vector of means and
returns the predicted parameters.}

\item{R2}{The goodness of fit \eqn{R^2} of the nonlinear fitting}

\item{df}{Degrees of freedom of nonlinear model. See \code{df} in
\code{\link[splines]{ns}} for more details.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
The final curve is fitted from \code{fit_log_trans_model}, so the parameters
was \eqn{log}-transformed first. \code{log_trans_NCS_plot} would plot 2
subplots. Subplot1 would draw mean-\eqn{log}-parameter in all sliding windows
and the final curve in \eqn{log} scale, subplot2 draws mean-parameter in all
sliding windows and the curve.
}
