% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{logL_fun}
\alias{logL_fun}
\title{Log-likelihood function}
\usage{
logL_fun(x, mj, tj)
}
\arguments{
\item{x}{A vector contains 2 parameters to be estimated.}

\item{mj}{A vector of M-value for a protein.}

\item{tj}{A vector of estimated technical variance for a protein.}
}
\value{
The value of log-likelihood function.
}
\description{
Log-likelihood function
}
