% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{logL_gr}
\alias{logL_gr}
\title{Gradient of log-likelihood function}
\usage{
logL_gr(x, mj, tj)
}
\arguments{
\item{x}{A numeric vector contains 2 parameters to be estimated.}

\item{mj}{A numeric vector of M-value for a protein.}

\item{tj}{A numeric vector of estimated technical variance for a protein.}
}
\value{
The gradient of log-likelihood function.
}
\description{
Gradient of log-likelihood function
}
