% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{logL_hessian}
\alias{logL_hessian}
\title{Hessian matrix of log-likelihood function}
\usage{
logL_hessian(x, mj, tj)
}
\arguments{
\item{x}{A vector contains 2 parameters to be estimated.}

\item{mj}{A vector of M-value for a protein.}

\item{tj}{A vector of estimated technical variance for a protein.}
}
\value{
The Hessian matrix of log-likelihood function.
}
\description{
Hessian matrix of log-likelihood function
}
