% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{MA_before_after_norm_plot}
\alias{MA_before_after_norm_plot}
\title{Draw the M-A plot in \eqn{log_2}-scale before and after normalization}
\usage{
MA_before_after_norm_plot(
  log2.int,
  log2.norm.int,
  sample_outdir,
  sample,
  ref,
  A.method
)
}
\arguments{
\item{log2.int}{A matrix with 2 columns. The first column is the \eqn{log_2}
intensity of current sample, and the second column is the \eqn{log_2}
intensity of its corresponding reference sample. They are not normalized.}

\item{log2.norm.int}{A matrix with 2 columns. The first column is the
\eqn{log_2}-intensity of current sample, and the second column is the
\eqn{log_2}-intensity of its corresponding reference sample. They are
normalized.}

\item{sample_outdir}{A character which represent the output directory.}

\item{sample}{The name of current sample to be fitted.}

\item{ref}{The name of reference sample related to current sample.}

\item{A.method}{A character string indicating the method to calculate
A-value. One of \code{"mean_of_2samples"}, \code{"ref"} or \code{"sample"}.
Usually A-value is regarded as the basic intensity of the observations and
it will take the intensity of reference.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
\code{MA_before_after_norm_plot} would draw 4 subplots. Subplot1 and subplot3
would draw \eqn{log_2}-intensity of the tissue sample against the reference
sample before and after normalization. Subplot2 and subplot4 would draw
M-value against A-value before and after normalization.
}
