% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalEngine.R
\name{MA_linear_regression}
\alias{MA_linear_regression}
\title{Fit linear model in current sliding window}
\usage{
MA_linear_regression(
  window.Num,
  groups,
  data,
  sample,
  middle.percent,
  PLOT,
  outdir
)
}
\arguments{
\item{window.Num}{The current number of sliding window.}

\item{groups}{A list of vectors for indexing observations. Each vector
effectively specifies a group of observations. \code{window.Num} together
with \code{groups} determines the index of observations which would be used
for linear model fitting.}

\item{data}{A matrix which records M-values and A-values, they are sorted by
the order of A-values. Those observations whose M-values and A-values are
\code{NA} will be removed.}

\item{sample}{The name of current sample to be fitted.}

\item{middle.percent}{A number which indicates which parameters should be
used for linear model fitting in the sliding window.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}
}
\value{
A vector records all of the results about linear model in current
  sliding window, which includes fitted variances, bias, mean of A-value in
  the window and goodness of fit \eqn{R^2}.
}
\description{
\code{window.Num} together with \code{groups} figures out the M-values used
for linear model fitting. \code{\link[stats]{lm}} models the linear
relationship between M-values and quantiles of standard normal distribution.
The slope of the linear model is local technical standard deviation of this
window and the intercept is the normalization bias.
}
