% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalPlot.R
\name{MA_sliding_plot}
\alias{MA_sliding_plot}
\title{Draw the MA-plot in each sliding window}
\usage{
MA_sliding_plot(
  A,
  M,
  windowA,
  windowM,
  outdir,
  sample,
  window.Num,
  fit.region,
  nonfit.region
)
}
\arguments{
\item{A}{A vector of total A-value, the basic intensity of the observations.}

\item{M}{A vector of total M-value, the \eqn{log_2} fold change.}

\item{windowA}{A vector of A-value in the given sliding window.}

\item{windowM}{A vector of M-value in the given sliding window.}

\item{outdir}{A character which represent the output directory.}

\item{sample}{The name of current sample to be fitted.}

\item{window.Num}{The current number of sliding window.}

\item{fit.region}{The region of proteins to be used for fitting the linear
model.}

\item{nonfit.region}{The region not used.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
\code{MA_sliding_plot} draw the MA-plot of all proteins. The X-axis is
\eqn{log_2}-intensity of reference sample, named A-value. The Y-axis is
\eqn{log_2} fold change, named M-value. To highlight the current fitted
window, proteins outside the window is in \code{"lightgrey"}. In the fitting
window, the region of proteins used for fitting are in \code{"red"} and
others are in \code{"grey"}.
}
