% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanTest.R
\name{meanTestPlot}
\alias{meanTestPlot}
\title{Plot a \code{meanTest} Object}
\usage{
meanTestPlot(
  res,
  ref.mean,
  P.cutoff = 0.05,
  p.type = c("pval", "padj"),
  mainAdd,
  raw_Diff = F,
  args.points = list(pch = 21, cex = 1.2),
  args.abline = list(col = "darkorange", lwd = 2.5, lty = 5),
  args.legend = list(x = "topright"),
  xlab = "mean",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{res}{A \code{meanTest} object.}

\item{ref.mean}{A numeric vector of mean among all reference samples.}

\item{P.cutoff}{A scalar between 0 and 1. Proteins with \emph{p}-values below
this cutoff are considered statistically significant under the null
hypothesis and are correspondingly highlighted in a distinct color in the
plot.}

\item{p.type}{A character which indicates the type of \emph{p}-value to show
in the plot. Must be one of \code{"pval"} (default), and \code{"padj"}. Can
be abbreviated.}

\item{mainAdd}{The added information to describe samples used to do test.}

\item{raw_Diff}{Whether to use the raw diff of m as y.}

\item{args.points}{A named list of graphical parameters for plotting the
point estimates (see \code{\link[graphics]{points}} for commonly used
graphical parameters).}

\item{args.abline}{A named list of graphical parameters for plotting the null
\emph{v_0} (see \code{\link[graphics]{abline}} for commonly used graphical
parameters).}

\item{args.legend}{A named list of graphical parameters for legend (see
\code{\link[graphics]{legend}} for commonly used graphical parameters).}

\item{xlab}{Label for the X axis.}

\item{xlim}{Range of X coordinate.}

\item{ylim}{Range of Y coordinate.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}}.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Given a \code{meanTest} object, which records the results of calling
HVPs and/or LVPs from a \code{\link{proObj}}, this method creates a scatter
plot of \code{(ref.mean, v)} pairs from all observations, marking
specifically the ones that are associated with statistical significance.
Besides, the theoretical \eqn{v_0} associated with this round of test of
\eqn{v} is added to the plot, serving as theoretical mean of \emph{v}.
}
\seealso{
\code{\link{proObj}} for creating a \code{proObj} object;
  \code{\link{estTechVar}} for fitting technical variance given a
  \code{proObj} objects; \code{\link{estBioVar}} for estimating biological
  mean \emph{m} and \emph{log}-transformed biological variance \emph{v};
  \code{varTest} for a test of hyper-variability.
}
