% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarPlot.R
\name{MeanVarPlot}
\alias{MeanVarPlot}
\title{Plot mean-\emph{v}}
\usage{
MeanVarPlot(
  ref.mean,
  v,
  Nsample,
  sample.type,
  group1,
  scaleV.raw.FC = NULL,
  outdir,
  outadd
)
}
\arguments{
\item{ref.mean}{A numeric vector of mean among all reference samples.}

\item{v}{A numeric vector of \eqn{log}-transformed biological variance
estimates \emph{v} for all proteins.}

\item{Nsample}{The sample size in this group.}

\item{sample.type}{The sample type of this group of samples.}

\item{group1}{A vector indicating the proteins in group1 according to the
classification of \emph{v}. If \code{scaleV.raw.FC} is not \code{NULL}, the
classification is by the lower bound \eqn{log(log_2^2(1+scaleV.raw.FC))}.}

\item{scaleV.raw.FC}{A specified average basic fluctuation around mean
expression of a protein. This parameter determines whether to set a lower
bound of \emph{v} and the specific setting of it.}

\item{outdir}{If \code{PLOT} is TRUE, the output directory.}

\item{outadd}{If \code{PLOT} is TRUE, the additions to the output documents.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
\code{FisherInfoPlot} would draw 2 pictures. The first one plot mean-\emph{v}
of all proteins. If \code{scaleV.raw.FC} is not \code{NULL}, the line of
lower bound is also added in the plot. The second one plot mean-\emph{v} of
proteins only in group1.
}
