% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{mkdir}
\alias{mkdir}
\title{Create new folders}
\usage{
mkdir(path_dir)
}
\arguments{
\item{path_dir}{A character declares the directory to create new folder.}
}
\value{
A newly created folder.
}
\description{
Create new folders according to the directory provided by users to restore
the process plots during the estimation.
}
