% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalEngine.R
\name{MS_quantile_position}
\alias{MS_quantile_position}
\title{Calculate the theoretical quantile of standard normal distribution}
\usage{
MS_quantile_position(start.pos, end.pos, aa)
}
\arguments{
\item{start.pos}{The start of theoretical plot position.}

\item{end.pos}{The end of theoretical plot position.}

\item{aa}{A scalar of scale parameter.}
}
\value{
The theoretical plot position in standard normal distribution.
}
\description{
According to the start number and end number of M-values, calculate their
theoretical quantile of standard normal distribution.
}
