% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSVar.R
\docType{package}
\name{MSVar}
\alias{MSVar-package}
\alias{MSVar}
\title{MSVar for Identifying Hypervariability proteins and Differentially Variable
proteins}
\description{
\code{MSVar} is designed for isobaric labeling-based mass spectrometry (ILMS)
quantitative proteomic profiling to identify hypervariable proteins for a
single group of samples and differentially variable proteins between
different groups via similar procedures.
}
\section{Maintainer}{
 Xuan Guo <\email{guoxuan2020@sinh.ac.cn}>
}

