% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobalEngine.R
\name{NCS_df_selection}
\alias{NCS_df_selection}
\title{Select a natural cubic spline model with high goodness of fit \eqn{R^2}}
\usage{
NCS_df_selection(x0, y0, choice)
}
\arguments{
\item{x0, y0}{Two numeric vectors of reference means and local variances in
the sliding window}

\item{choice}{A numeric to specify how many natural cubic spline nonlinear
models to fit to choose from. By default, 2 non linear models with
\code{df} in \code{\link[splines]{ns}} would be fitted and the one with
higher goodness of fit \eqn{R^2} would be served as the final nonlinear
model.}
}
\value{
A list which contains the final nonlinear model, goodness of fit
  \eqn{R^2} and \code{df}.
}
\description{
Select a natural cubic spline model with high goodness of fit \eqn{R^2}
}
