% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{Newton_optimize}
\alias{Newton_optimize}
\title{Nonlinear minimization of log-likelihood function}
\usage{
Newton_optimize(MValue, EstTechVar)
}
\arguments{
\item{MValue}{A vector of M-value for a protein.}

\item{EstTechVar}{A vector of estimated technical variance for a protein.}
}
\value{
A list contains the maximum likelihood estimates for
  \eqn{log}-transformed biological variance and biological mean, along with
  the corresponding minimized value of the log-likelihood function.
}
\description{
Use Newton-iteration to minimize the log-likelihood function for the
parameter of biological variance and mean.
}
