% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalize the intensity for each sample against the reference sample}
\usage{
normalize(
  raw.intensity,
  filter.low = T,
  offset = 1,
  L = 1.5,
  A.method = c("mean_of_2samples", "ref", "sample"),
  PLOT = FALSE,
  outdir
)
}
\arguments{
\item{raw.intensity}{A raw intensity matrix which is composed by a sample and
the corresponding reference sample.}

\item{filter.low}{Whether to filter proteins whose raw intensity is under 10?
The default is TRUE.}

\item{offset}{A scalar used as an offset added to the raw intensity, avoiding
infinity while \eqn{log_2}-transformation. Default is 1.}

\item{L}{A scalar to scale the cutoff of outliers' boundary.}

\item{A.method}{A character string indicating the method to calculate
A-value. One of \code{"mean_of_2samples"}, \code{"ref"} or \code{"sample"}.
Usually A-value is regarded as the basic intensity of the observations and
it will take the intensity of reference.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the directory of output figures.}
}
\value{
A normalized intensity 3-column matrix with normalized intensity,
  M-value and A-value in the \eqn{log_2}-scale.
}
\description{
Given the raw intensity for each sample and the corresponding reference
sample, conduct 1v1 normalization. Only use proteins which are non-NA in both
samples to calculate trimmed total intensity, and then obtain the
normalization factor for each sample. For all proteins in the sample, its raw
intensity would divide the normalization factor to get the normalized
intensity. However, the reference sample remain unchanged.
}
\details{
After normalization, intensity in each sample and reference sample would take
\eqn{log_2}-transformation and \code{normalize} would calculate A-values and
M-values in \eqn{log_2}-scale. A-value is regarded as the basic intensity of
the observations, and M-value is \eqn{log_2}-transformed fold change.
}
