% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorM.R
\name{PostM}
\alias{PostM}
\title{Derive posterior M-value}
\usage{
PostM(x, scaleV.raw.FC = 0.1)
}
\arguments{
\item{x}{A \code{\link{proObj}}.}

\item{scaleV.raw.FC}{A specified average basic fluctuation around mean
expression of a protein. The default setting of \code{scaleV.raw.FC} is
10\%, which amounts to an average fluctuation of 10\% around mean expression.
Accordingly, the baseline fluctuation for each protein is defined as
\eqn{log(log_2(1+10\%)^2)}.}
}
\value{
A \code{\link{proObj}} with posterior M-value.
}
\description{
MSVar derives a posterior estimate of each \eqn{M'_ij} by finding the maximum
a posteriori.
}
