% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proObj.R
\name{print.proObj}
\alias{print.proObj}
\title{Print a \code{proObj}}
\usage{
\method{print}{proObj}(x, ...)
}
\arguments{
\item{x}{A \code{\link{proObj}}.}

\item{...}{Arguments passed from other methods.}
}
\value{
The function returns \code{x} invisibly.
}
\description{
This function prints its argument, which is a \code{\link{proObj}}, and
returns it invisibly (via \code{\link[base]{invisible}(x)}).
}
\details{
This function implements the \code{\link[base]{print}} method for the
\code{"\link{proObj}"} class.
}
\seealso{
\code{\link{proObj}} for creating a \code{proObj}.
}
