% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proObj.R
\name{proObj}
\alias{proObj}
\title{Create a \code{proObj}}
\usage{
proObj(
  sample.intensity,
  ref.intensity,
  sample.ref.cor,
  IDs = NULL,
  name = "NA"
)
}
\arguments{
\item{sample.intensity}{A numeric matrix of a group of samples' raw
intensity.}

\item{ref.intensity}{A numeric matrix of all reference samples' raw
intensity.}

\item{sample.ref.cor}{A named character vector which records the
corresponding reference sample name of each tissue sample in
\code{sample.intensity}.}

\item{IDs}{A character vector of proteins' names.}

\item{name}{A character scalar specifying the name of this \code{proObj}.
Used only for demonstration.}
}
\value{
A newly constructed object of the \code{\link[base]{class}}
  \code{"proObj"}.
}
\description{
This function constructs a new \code{proObj} based on the given raw proteins
intensity of a group of samples and reference samples, and the corresponding
relationship is indicated by \code{sample.ref.cor}.
}
\seealso{
\code{\link{proObjFromTMT}} for creating a \code{proObj} from
  general result from TMT experiment.
}
