% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proObj.R
\name{proObjFromTMT}
\alias{proObjFromTMT}
\title{Create a \code{proObj}}
\usage{
proObjFromTMT(raw.intensity, batch.info, IDs = NULL, name = "NA")
}
\arguments{
\item{raw.intensity}{A numeric matrix of raw intensity, which includes a
group of samples and reference samples.}

\item{batch.info}{A character matrix of batch information, which indicates
the corresponding relationship of a group of samples among batches. Each
row includes the reference sample name and sample names in that batch.}

\item{IDs}{A character vector of proteins' names.}

\item{name}{A character scalar specifying the name of this \code{proObj}.
Used only for demonstration.}
}
\value{
A newly constructed object of the \code{\link[base]{class}}
  \code{"proObj"}.
}
\description{
This function constructs a new \code{proObj} based on the general TMT
experiment result, which include raw intensity of all samples and batch
information of the MS experiment.
}
\seealso{
\code{\link{proObj}} for another constructor of \code{proObj}s.
}
