% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarLocalPlot.R
\name{r2_linear_boxplot}
\alias{r2_linear_boxplot}
\title{Draw a \code{boxplot} of \eqn{R^2} of linear model for each sample}
\usage{
r2_linear_boxplot(linear_r2, outdir, sample, ylim = c(0.95, 1))
}
\arguments{
\item{linear_r2}{A numeric vector of \eqn{R^2} in all sliding windows.}

\item{outdir}{A character which represent the output directory.}

\item{sample}{The name of current fitted sample.}

\item{ylim}{Range of Y coordinate.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Draw a \code{boxplot} of \eqn{R^2} of linear model for each sample
}
