% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBioMeanVarEngine.R
\name{resBioCMB}
\alias{resBioCMB}
\title{Combine the result of biological variance estimation}
\usage{
resBioCMB(
  scaleV.raw.FC = NULL,
  ref.mean,
  m,
  v,
  tech.var,
  Nsample,
  sample.type,
  pro.name,
  PLOT,
  outdir,
  outadd
)
}
\arguments{
\item{scaleV.raw.FC}{A specified average basic fluctuation around mean
expression of a protein. This parameter determines whether to set a lower
bound of \emph{v} and the specific setting of it.}

\item{ref.mean}{A numeric vector of mean among all reference samples.}

\item{m}{A numeric vector of biological mean estimates \emph{m} for all
proteins.}

\item{v}{A numeric vector of \eqn{log}-transformed biological variance
estimates \emph{v} for all proteins.}

\item{tech.var}{A numeric matrix of technical variance of all proteins in all
samples. The rows are proteins and the columns are samples.}

\item{Nsample}{The sample size in this group.}

\item{sample.type}{The sample type of this group of samples.}

\item{pro.name}{A character vector of all proteins' names.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is TRUE, the output directory.}

\item{outadd}{If \code{PLOT} is TRUE, the additions to the output documents.}
}
\value{
A combined matrix with estimated \emph{m} & \emph{v} value and their
  fisher information matrix.
}
\description{
Combine the result of biological variance estimation
}
