% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTechVarGlobalEngine.R
\name{Var_NonLinear_Fit}
\alias{Var_NonLinear_Fit}
\title{\code{Var_NonLinear_Fit} returns estimated technical variance}
\usage{
Var_NonLinear_Fit(
  para.fit.method,
  windowMean,
  windowVar,
  used.global.fit,
  choice = 2,
  args.locfit = list(),
  AValue,
  PLOT = F,
  outdir = "./",
  sample
)
}
\arguments{
\item{para.fit.method}{A character string indicating the method to be used
for fitting the variance curve globally. One of
\code{"natural_cubic_spline"}, \code{"log_trans"} (default) or
\code{"local_regression"}}

\item{windowMean}{A vector of mean A-value in all sliding windows.}

\item{windowVar}{A vector of estimated local variance in all sliding windows.}

\item{used.global.fit}{A vector of 2 with each element between 0 and 1
specifying a subset of sliding windows as the input of the whole globally
curve fitting process. Since the locally estimated parameters in first and
end parts of the sliding windows may be unstable and unreliable,
\code{used.global.fit} can specify that only this part of the sliding
windows are used. The default is \code{c(0.05, 0.95)}.}

\item{choice}{If \code{para.fit.method} is \code{"natural_cubic_spline"},
\code{"log_trans"}, \code{choice} is a numeric to specify how many natural
cubic spline nonlinear models to fit to choose from. By default, 2 non
linear models with \code{df} in \code{\link[splines]{ns}} would be fitted
and the one with higher goodness of fit \eqn{R^2} would be served as the
final nonlinear model. Otherwise, if \code{para.fit.method} is
\code{"local_regression"}, \code{choice} is a numeric to specify how many
local regression nonlinear models to fit to choose from. By default, 2 non
linear models with \code{nn} in \code{\link[locfit]{lp}} would be fitted
and the one with higher goodness of fit \eqn{R^2} would be served as the
final nonlinear model.}

\item{args.locfit}{A named list of extra arguments to
\code{\link[locfit]{locfit}}. Only used when \code{method} is
\code{"local"}.}

\item{AValue}{A vector of A-value. \code{AValue} would be used to predict the
technical variance from the nonlinear variance prediction function. If
\code{AValue} is more than or less than the boundary of \code{windowMean}
which were used for nonlinear model fitting, the predictions at the
boundary would be set as the estimated variance at the endpoints.}

\item{PLOT}{Whether to draw the process plots?}

\item{outdir}{If \code{PLOT} is true, the output directory.}

\item{sample}{The name of current sample to be fitted.}
}
\value{
A list include globally fitted technical variance prediction
  function and the final estimated technical variance.
}
\description{
Technical variance is a function of reference for each sample, nonlinear
models would be fitted to describe the relationship between them. By using
\code{estTechVarLocal}, technical variance in each sliding window are
obtained and stored as \code{windowVar} and \code{windowMean}. Since the
locally estimated parameters in first and end parts of the sliding windows
may be unstable and unreliable, \code{used.global.fit} can specify that only
this part of the sliding windows are used during the curve fitting process.
}
\details{
While fitting the nonlinear curve, there are 3 methods to choose.
\code{"natural_cubic_spline"} and \code{"log_trans"} actually use natural
cubic spline, but \code{"local_regression"} uses local regression. No matter
which method is used, 2 or 3 models would be fitted depends on \code{choice}
and the one with higher goodness of fit \eqn{R^2} is the final nonlinear
model.

After fitting the curve, \code{AValue} is used as input of prediction of
technical variance and more judgement need to be done to decide whether use
estimation from the curve directly. If \code{AValue} is more than or less
than the boundary of \code{windowMean} which were used for nonlinear model
fitting, the positive predictions at the boundary would be set as the
corresponding estimated variance. Besides, if the goodness of fit \eqn{R^2}
of this nonlinear curve is less than 0.5, mean of \code{windowVar} would be
set as estimated variance for all observations.
}
