% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarTest.R
\name{VarTest}
\alias{VarTest}
\title{Identify hypervariable proteins}
\usage{
VarTest(x, null.v0 = 0.2, p.adjust.method = "BH")
}
\arguments{
\item{x}{A \code{\link{proObj}} with estimated \emph{v} and fisher
information.}

\item{null.v0}{A pre-specified constant \eqn{v_0} used for HVP
identification. The default setting of \eqn{v_0} amounts to an average
fluctuation of 20\% around mean expression, that is, the standard deviation
at the scale of M-values is \eqn{log_2 (1+20\%)} and \eqn{v_0 =
log(log_2(1+20\%)^2)}.}

\item{p.adjust.method}{The method for adjusting \emph{p}-values for multiple
testing. See \code{\link[stats]{p.adjust}} for all options.}
}
\value{
An object of the \code{\link[base]{class}} \code{c("varTest",
  "data.frame")}, which records the hyper variability test results for each
  region by each row.
}
\description{
Given a \code{\link{proObj}}, this function identifies hypervariable proteins
(HVPs) by performing a statistical test for each protein, with the resulting
\emph{p}-value assessing whether the observed \emph{log}-transformed variance
\emph{v} is significantly larger than a pre-specified constant.
}
\details{
Proteins which are located in the lower part of \emph{v} are obviously
non-HVP, their \emph{p}-values are set as \code{NA} directly.
}
